+++ common/src/main/kotlin/me/znotchill/marmot/common/ui/components/TextComponent.kt
@@ -13,13 +13,13 @@ open class TextComponent(
 ) : Component() {
     override val compType: CompType = CompType.TEXT
 
-    override fun width(): Int {
+    override fun width(): Float {
         val baseWidth = props.text.mcWidth() + props.padding.left + props.padding.right
-        return (baseWidth * props.scale.x).toInt()
+        return (baseWidth * props.scale.x)
     }
 
-    override fun height(): Int {
+    override fun height(): Float {
         val baseHeight = 7 + props.padding.top + props.padding.bottom
-        return (baseHeight * props.scale.y).toInt()
+        return (baseHeight * props.scale.y)
     }
 }
